analyzer/requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# LLM service APIs
openai>=1.0.0          # For GPT models
google-generativeai>=0.3.0  # For Gemini
# For Llama via Groq
groq>=0.4.0
# OR for Llama via Ollama
ollama>=0.1.0

# Image processing
Pillow>=9.5.0
base64io>=1.0.3
